+++ product_code/course_service/impl/src/main/scala/de/upb/cs/uc4/course/impl/CourseServiceImpl.scala
@@ -6,29 +6,30 @@ import akka.{ Done, NotUsed }
 import com.fasterxml.uuid.Generators
 import com.lightbend.lagom.scaladsl.api.ServiceCall
 import com.lightbend.lagom.scaladsl.api.transport._
-import com.lightbend.lagom.scaladsl.persistence.ReadSide
 import com.lightbend.lagom.scaladsl.server.ServerServiceCall
 import com.typesafe.config.Config
 import de.upb.cs.uc4.authentication.model.AuthenticationRole
 import de.upb.cs.uc4.course.api.CourseService
 import de.upb.cs.uc4.course.impl.actor.CourseState
 import de.upb.cs.uc4.course.impl.commands._
-import de.upb.cs.uc4.course.impl.readside.{ CourseDatabase, CourseEventProcessor }
+import de.upb.cs.uc4.course.impl.readside.CourseDatabase
 import de.upb.cs.uc4.course.model.Course
 import de.upb.cs.uc4.shared.client.exceptions._
 import de.upb.cs.uc4.shared.server.ServiceCallFactory._
-import de.upb.cs.uc4.shared.server.messages.{ Accepted, Confirmation, Rejected, RejectedWithError }
+import de.upb.cs.uc4.shared.server.messages.{ Accepted, Confirmation, Rejected }
 import de.upb.cs.uc4.user.api.UserService
+import play.api.Environment
 
 import scala.concurrent.duration._
 import scala.concurrent.{ Await, ExecutionContext, Future, TimeoutException }
 
 /** Implementation of the CourseService */
 class CourseServiceImpl(
-    clusterSharding: ClusterSharding, userService: UserService,
-    readSide: ReadSide, processor: CourseEventProcessor, database: CourseDatabase
+    clusterSharding: ClusterSharding,
+    userService: UserService,
+    database: CourseDatabase,
+    override val environment: Environment
 )(implicit ec: ExecutionContext, config: Config) extends CourseService {
-  readSide.register(processor)
 
   /** Looks up the entity for the given ID */
   private def entityRef(id: String): EntityRef[CourseCommand] =
@@ -86,7 +87,7 @@ class CourseServiceImpl(
           // Check if the lecturer does exist
           userService.getUser(courseProposal.lecturerId).handleRequestHeader(addAuthenticationHeader(header)).invoke().recover {
             //If the lecturer does not exist, we throw a validation error containing that info
-            case ex: UC4Exception if ex.errorCode.http == 404 =>
+            case ex: UC4Exception if ex.errorCode == 404 =>
               throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationErrors :+ SimpleError("lecturerId", "Lecturer does not exist")))
           }.flatMap { _ =>
             if (validationErrors.nonEmpty) {
@@ -99,9 +100,9 @@ class CourseServiceImpl(
 
             ref.ask[Confirmation](replyTo => CreateCourse(courseToAdd, replyTo))
               .map {
-                case Accepted => // Creation Successful
+                case Accepted(_) => // Creation Successful
                   (ResponseHeader(201, MessageProtocol.empty, List(("Location", s"$pathPrefix/courses/${courseToAdd.courseId}"))), courseToAdd)
-                case RejectedWithError(code, reason) =>
+                case Rejected(code, reason) =>
                   throw UC4Exception(code, reason)
               }
           }
@@ -122,10 +123,10 @@ class CourseServiceImpl(
               else {
                 entityRef(id).ask[Confirmation](replyTo => DeleteCourse(id, replyTo))
                   .map {
-                    case Accepted => // OK
+                    case Accepted(_) => // OK
                       (ResponseHeader(200, MessageProtocol.empty, List()), Done)
-                    case Rejected(reason) => // Not Found
-                      throw UC4Exception.InternalServerError("Course Deletion Error", reason)
+                    case Rejected(code, reason) => // Not Found
+                      throw UC4Exception(code, reason)
                   }
               }
             case None =>
@@ -169,7 +170,7 @@ class CourseServiceImpl(
 
         // Check if the lecturer does exist
         userService.getUser(updatedCourse.lecturerId).handleRequestHeader(addAuthenticationHeader(header)).invoke().recover {
-          case ex: UC4Exception if ex.errorCode.http == 404 =>
+          case ex: UC4Exception if ex.errorCode == 404 =>
             throw new UC4NonCriticalException(422, DetailedError(ErrorType.Validation, validationErrors :+ SimpleError("lecturerId", "Lecturer does not exist.")))
         }.flatMap { _ =>
           val ref = entityRef(id)
@@ -186,9 +187,9 @@ class CourseServiceImpl(
               else {
                 ref.ask[Confirmation](replyTo => UpdateCourse(updatedCourse, replyTo))
                   .map {
-                    case Accepted => // Update Successful
+                    case Accepted(_) => // Update Successful
                       (ResponseHeader(200, MessageProtocol.empty, List()), Done)
-                    case RejectedWithError(code, reason) =>
+                    case Rejected(code, reason) =>
                       throw UC4Exception(code, reason)
                   }
               }