+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/SignedTransactionProposal.scala
@@ -2,7 +2,6 @@ package de.upb.cs.uc4.shared.client
 
 import java.util.Base64
 
-import com.google.protobuf.ByteString
 import play.api.libs.json.{ Format, Json }
 
 case class SignedTransactionProposal(unsignedProposal: String, signature: String) {