+++ product_code/configuration_service/CHANGELOG.md
@@ -1,8 +1,21 @@
-# [WIP v0.11.1](https://github.com/upb-uc4/University-Credits-4.0/compare/configuration-v0.11.0...configuration-v0.11.1) (2020-XX-XX)
+# [v0.12.1](https://github.com/upb-uc4/University-Credits-4.0/compare/configuration-v0.12.0...configuration-v0.12.1) (2020-11-11)
+## Feature
+ - Added error messages for every regex in ValidationConfiguration
+## Refactor
+## Bugfix
+
+# [v0.12.0](https://github.com/upb-uc4/University-Credits-4.0/compare/configuration-v0.11.1...configuration-v0.12.0) (2020-11-10)
+## Feature
+## Refactor
+ - Bumped Version of Lagom to 1.6.4
+## Bugfix
+
+# [v0.11.1](https://github.com/upb-uc4/University-Credits-4.0/compare/configuration-v0.11.0...configuration-v0.11.1) (2020-11-06)
 ## Feature
  - Added support for ETags
 ## Refactor
 ## Bugfix
+ - Added all messages to serialization registry
 
 # [v0.11.0](https://github.com/upb-uc4/University-Credits-4.0/compare/configuration-v0.10.1...configuration-v0.11.0) (2020-10-26)
 ## Feature