+++ product_code/course_service/CHANGELOG.md
@@ -1,8 +1,15 @@
-# [WIP v0.11.1](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.11.0...course-v0.11.1) (2020-XX-XX)
+# [v0.12.0](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.11.1...course-v0.12.0) (2020-11-10)
+## Feature
+## Refactor
+ - Bumped Version of Lagom to 1.6.4
+## Bugfix
+
+# [v0.11.1](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.11.0...course-v0.11.1) (2020-11-06)
 ## Feature
  - Added support for ETags
 ## Refactor
 ## Bugfix
+ - Added all messages to serialization registry
 
 # [v0.11.0](https://github.com/upb-uc4/University-Credits-4.0/compare/course-v0.10.0...course-v0.11.0) (2020-10-26)
 ## Feature